Simplify linter, using a strategy method.
